requirements_lint.txt
# Lint and static analysis dependencies.
tox==4.30.2
flake8==7.3.0
flake8-bugbear==24.12.12
flake8-comprehensions==3.16.0
flake8-pytest-style==2.1.0
mypy==1.18.2
pylint==3.3.9
pymarkdownlnt==0.9.31
yamllint==1.37.1
